move GameBoardBuilder to builders.hpp
