+++ cmake/add_subdirectories.cmake
@@ -1,3 +1,4 @@
+add_subdirectory(${SOURCE_CODE_DIR}/builders)
 add_subdirectory(${SOURCE_CODE_DIR}/game_board)
 add_subdirectory(${SOURCE_CODE_DIR}/game_board_for_concepts)
