+++ src/core/builders/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(builders builders.cpp)
+
+target_link_libraries(builders PUBLIC game_board_for_concepts)
