+++ src/core/game_board_for_concepts/game_board_for_concepts.cpp
@@ -180,10 +180,5 @@ bool GameBoardForConcepts::ViolatesRepeatRule(PieceColor color) {
   return false;
 }
 
-std::shared_ptr<GameBoardForConcepts> GameBoardBuilder::build(
-    const BoardMapInt_t starting_board
-) {
-  return std::make_shared<GameBoardForConcepts>(starting_board);
-}
 
 } // namespace gameboard
